update following 2nd review
Student correctly describes the k-fold cross-validation technique and discusses the benefits of its application when used with grid search when optimizing a model.
Required

What you said in the first paragraph is perfect:

The k-fold cross-validation is a training technique that randomly splits the training dataset into k folds without replacement, where(k-1) folds are used for the model training and one fold is used for testing. This procedure is repeated k times so that we obtain k models and performance estimates. Since k-fold cross-validation is a resampling technique without replacement it takes into account variance in the training set. The main idea here is that each sample in the training set has the opportunity of being tested (or evaluated). ...
However, you need to review the continuation:

... The model with the best performance is selected from the k models. ...
Actually, all the performance estimates are averaged, as illustrated in the following image: